create component for eventName, billOwner, and date
This was achieved by lifting the state and adding an udpateBill function
to BillUserInterface
Also a preliminary step to create multiple events
This commit has those beginning stages as well, such as a Add Icon tab for
adding new events
Still need to figure out a wise place to update the tabs per Bill
